move oauth code and api handler under pro
